Test that duplicate transactions don't debit/credit more than once Test that duplicate transactions and that duplicate OTS index doesn't get accepted (but fails currently because CLI tx_sign doesn't update the wallet's ots index) DRY refactoring
